+++ tests/base_tests.py
@@ -1556,7 +1556,6 @@ class GetFeatureNamesOutTests:
 
         if x.FITS:
             x.fit(df, df["a"])
-
         output = x.transform(df)
 
         output_columns = set(output.columns)