+++ tests/conftest.py
@@ -210,6 +210,7 @@ def minimal_attribute_dict():
         "InteractionTransformer": {
             "columns": ["a", "b"],
         },
+        "LowerCaseTransformer": {"columns": ["b"]},
         "LogTransformer": {
             "columns": ["a"],
         },
@@ -261,6 +262,10 @@ def minimal_attribute_dict():
         },
         "PCATransformer": {
             "columns": ["a", "c"],
+        },
+        "RemoveCharactersTransformer": {
+            "columns": ["b"],
+            "characters": [r"\s"],
         },
         "RenameColumnsTransformer": {
             "columns": "a",