+++ CHANGELOG.rst
@@ -27,11 +27,13 @@ We use the tags:
 
 Each individual change should have a link to the pull request after the description of the change.
 
-3.4.0 (placeholder)
+3.4.0 (unreleased)
 ------------------
 
 Changed
 ^^^^^^^
+- added LowerCaseTransformer and RemoveCharactersTransformer to string module
+- added functions module for stateless transforms
 - removed narwhals private import from nominal module `#725 <https://github.com/azukds/tubular/issues/725>_`
 
 3.3.0 (30/04/2026)